requirements.txt
pyglet~=2.0.20
pillow~=11.0.0